requirements.txt
requests
numpy
faiss-cpu
sentence_transformers
python-dotenv